Updated README, hatching
This updates the README to account for changes since we updated the API
tables as well as set the Swagger UI into place.

This is still a pilot and work in progress.

Started to explore a `hatch` package for a library that might serve the
community more easily. Again, a very early exploration.